+++ app/models/close_encounters/participant_event.rb
@@ -6,5 +6,9 @@ module CloseEncounters
       foreign_key: "close_encounters_participant_service_id"
 
     scope :newest, -> { order(created_at: :desc).limit(1) }
+
+    unless ActiveRecord::Base.connection.adapter_name.downcase.include?("postgresql")
+      serialize :metadata, coder: JSON
+    end
   end
 end