+++ lib/close_encounters.rb
@@ -19,6 +19,23 @@ module CloseEncounters
     end
   end
 
+  # Record a verification of a contact with a third party service where the
+  # verification is a callable which must also respond to to_s.
+  #
+  # For example, provide a callable which checks the JSON Schema for a response body
+  # and will record an event if calling the verification returns false.
+  #
+  # @param name [String] the name of the service
+  # @param status [Integer] the HTTP status of the contact
+  # @param response [String] the response object
+  # @param verifier [Proc] the verification callable which must also respond to to_s
+  def verify(name, status:, response:, verifier:)
+    service = ParticipantService.find_by!(name:)
+    unless service.events.newest.pick(:status) == status && (verified = verifier.call(response))
+      service.events.create!(status:, response:, metadata: {verified:, verification: verifier.to_s})
+    end
+  end
+
   # Determine if contacts with third party services should be recorded automatically
   # using the Rack Middleware
   #